requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
arxiv==2.1.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
etelemetry==0.3.1
fastapi==0.115.3
feedparser==6.0.11
ffmpy==0.4.0
filelock==3.16.1
fitz==0.0.1.dev2
frontend==0.0.3
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
gradio==5.3.0
gradio_client==1.4.2
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nest-asyncio==1.6.0
networkx==3.4.1
nibabel==5.3.1
nipype==1.8.6
numpy==2.1.2
orjson==3.10.9
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
prov==2.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.12
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytils==0.4.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
requests==2.32.3
rich==13.9.3
rsa==4.9
ruff==0.7.0
scipy==1.14.1
semantic-version==2.10.0
semanticscholar==0.8.4
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
tenacity==9.0.0
toml==0.10.2
tomlkit==0.12.0
tools==0.1.9
tqdm==4.66.5
traits==6.3.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
